Split nc_xcor.c into its three tiers
nc_xcor.c held the object, the redshift-space Limber methods and the whole
outer k quadrature in 2359 lines. The two tiers share no code -- one integrates
over z with k pinned to (l + 1/2) / chi(z), the other integrates over k with a
pair of fitted closures -- so they are now separate translation units,
nc_xcor_limber_z.c and nc_xcor_kquad.c. What stays behind is the object, its
properties, nc_xcor_compute(), the tier dispatch and the Limber-disjoint
policy: 849 lines.

Pure code motion. The five functions the split promoted from static are
declared in nc_xcor_priv.h, which now also carries struct _NcXcor and
NcXcorArg, and _nc_xcor_check_qag_status() stays in nc_xcor.c because both
tiers use it. Nothing else changed: the sweep over the full k-integral case
matrix -- 7344 rows, 17 pairs, 2 closures, 3 methods -- reproduces every C_ell
bit for bit against master.
